Refactor : jwt 코드 수정
