+++ matched/src/main/java/com/linked/matched/controller/user/UserController.java
@@ -1,6 +1,6 @@
 package com.linked.matched.controller.user;
 
-import com.linked.matched.request.jwt.TokenDto;
+import com.linked.matched.response.jwt.TokenDto;
 import com.linked.matched.request.user.UserJoin;
 import com.linked.matched.request.user.UserLogin;
 import com.linked.matched.service.user.UserService;
